requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
captcha==0.6.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==30.4.0
favicon==0.7.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.0
h5py==3.12.1
htbuilder==0.6.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
namex==0.0.8
narwhals==1.9.3
networkx==3.4.1
nltk==3.9.1
numpy==1.26.4
openai==0.28.0
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
plotly==5.24.1
preshed==3.0.9
prettytable==3.11.0
prometheus_client==0.21.0
propcache==0.2.0
protobuf==4.25.5
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyecharts==2.0.6
Pygments==2.18.0
PyJWT==2.9.0
pymdown-extensions==10.11.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
sqlitecloud==0.0.83
srsly==2.4.8
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.41.0
streamlit-authenticator==0.3.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-plotly-events==0.0.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
#tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
#torchaudio==2.5.0+cpu
#torchvision==0.20.0+cpu
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
wkhtmltopdf==0.2
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.15.3
streamlit-chat 
google-generativeai
